Refactored to move methods to location fitting general scheme.
Also added get_templates
